# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/create_convo": {
      "post": {
        "operationId": "GroupsController_create_convo",
        "summary": "Create personal conversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConvoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created convo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "User_Doesnt_Exist": {
                    "value": {
                      "statusCode": 400,
                      "message": "User does not exists",
                      "error": "Bad Request"
                    }
                  },
                  "Already_Exists": {
                    "value": {
                      "statusCode": 400,
                      "message": "Conv already exists",
                      "error": "Bad Request"
                    }
                  },
                  "No_Self_Conv": {
                    "value": {
                      "statusCode": 400,
                      "message": "You can't create a conversation with yourself",
                      "error": "Bad Request"
                    }
                  },
                  "More_than_2_users": {
                    "value": {
                      "statusCode": 400,
                      "message": "More than 2 users in this conversation",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/create_group": {
      "post": {
        "operationId": "GroupsController_create_group",
        "summary": "Create group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created Group"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "User_Doesnt_Exist": {
                    "value": {
                      "statusCode": 400,
                      "message": "User does not exists",
                      "error": "Bad Request"
                    }
                  },
                  "Cannot_Upload_Image": {
                    "value": {
                      "statusCode": 400,
                      "message": "Cannot upload image",
                      "error": "Bad Request"
                    }
                  },
                  "Duplicate_Entries": {
                    "value": {
                      "statusCode": 400,
                      "message": "Duplicate entries",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/get_chats": {
      "get": {
        "operationId": "GroupsController_get_all_chats",
        "summary": "Get all groups for a user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "chats": [
                      {
                        "_id": "string",
                        "name": "Super cool group",
                        "groupType": [
                          "GROUP"
                        ],
                        "latestMessage": {
                          "_id": "string",
                          "send_at": "2023-03-20T19:44:19.883Z",
                          "user_id": "string",
                          "text": "Hello!"
                        }
                      },
                      {
                        "_id": "string",
                        "name": "ADHD",
                        "groupType": [
                          "GROUP"
                        ],
                        "groupImage": "string",
                        "latestMessage": {
                          "_id": "string",
                          "send_at": "2023-04-04T08:19:08.508Z",
                          "user_id": "string",
                          "text": "testing time!"
                        }
                      },
                      {
                        "_id": "string",
                        "name": "Jad Zarzour",
                        "groupType": [
                          "PRIVATE"
                        ],
                        "groupImage": "string",
                        "latestMessage": {
                          "_id": "string",
                          "send_at": "2023-04-13T05:53:33.190Z",
                          "user_id": "string",
                          "text": "hi"
                        }
                      },
                      {
                        "_id": "string",
                        "name": "Very cool",
                        "groupImage": "string",
                        "groupType": [
                          "GROUP"
                        ]
                      },
                      {
                        "_id": "string",
                        "name": "string string",
                        "groupType": [
                          "PRIVATE"
                        ]
                      },
                      {
                        "_id": "string",
                        "name": "string string",
                        "groupImage": null,
                        "groupType": [
                          "PRIVATE"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/user_convo": {
      "get": {
        "operationId": "GroupsController_get_users_to_create_convo",
        "summary": "Gets all the users that a user can create a convo with",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "users": [
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "Jhon",
                        "lastName": "Doe"
                      },
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "Mart",
                        "lastName": "Slavin"
                      },
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "string",
                        "lastName": "string"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/user_group": {
      "get": {
        "operationId": "GroupsController_get_users_to_create_group",
        "summary": "Gets all the users that a doctor can create a group with",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "users": [
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "Jhon",
                        "lastName": "Doe"
                      },
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "Mart",
                        "lastName": "Slavin"
                      },
                      {
                        "_id": "string",
                        "email": "string",
                        "firstName": "string",
                        "lastName": "string"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/users/{group_id}": {
      "get": {
        "operationId": "GroupsController_get_users",
        "parameters": [
          {
            "name": "group_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Theraline endpoints",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConvoDto": {
        "type": "object",
        "properties": {
          "users_id": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "users_id"
        ]
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "users_id": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "users_id",
          "name"
        ]
      },
...
    }
  }
}